Fixed small bugs with pick list and rankings

+++ app/src/main/java/frc3824/rohawkticsscouting2017/Firebase/DataModels/TeamRankingData.java
@@ -17,7 +17,7 @@ public class TeamRankingData {
 
     // Ranking Data
     public int rank;
-    public int RPs;
+    public float RPs;
     public int wins;
     public int ties;
     public int losses;
